+++ internal/news/section.go
@@ -1,0 +1,12 @@
+package news
+
+const (
+	LatestSection      = "latest_section"
+	EditorPicksSection = "editor_picks_section"
+	LatestTopicSection = "latest_topic_section"
+	ReviewsSection     = "reviews_section"
+	CategoriesSection  = "categories_posts_section"
+	TopicsSection      = "topics_section"
+	PhotoSection       = "photos_section"
+	InfographicSection = "infographics_section"
+)