+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 	"os"
@@ -9,10 +10,13 @@ import (
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	f "github.com/twreporter/logformatter"
+	"go.mongodb.org/mongo-driver/mongo/options"
+	"go.mongodb.org/mongo-driver/mongo/readpref"
 
 	"twreporter.org/go-api/configs"
 	"twreporter.org/go-api/controllers"
 	"twreporter.org/go-api/globals"
+	"twreporter.org/go-api/internal/mongo"
 	"twreporter.org/go-api/routers"
 	"twreporter.org/go-api/services"
 	"twreporter.org/go-api/utils"
@@ -55,10 +59,21 @@ func main() {
 		return
 	}
 
+	log.Info("Connection to MongoDB with mongo-go-driver")
+	ctx := context.Background()
+	opts := options.Client()
+	client, err := mongo.NewClient(ctx, opts.ApplyURI(globals.Conf.DB.Mongo.URL).SetReadPreference(readpref.Nearest()))
+
+	if err != nil {
+		return
+	}
+	defer func() {
+		client.Disconnect(ctx)
+	}()
 	// mailSender := services.NewSMTPMailService() // use office365 to send mails
 	mailSvc := services.NewAmazonMailService() // use Amazon SES to send mails
 
-	cf = controllers.NewControllerFactory(db, session, mailSvc)
+	cf = controllers.NewControllerFactory(db, session, mailSvc, client)
 
 	// set up the router
 	router := routers.SetupRouter(cf)